CI and unit-testing for TB

+++ libCuik/meta/lexgen.c
@@ -183,7 +183,12 @@ static int ns(void) { return table_id_counter++; }
 
 #define RANGE(old, new, ...) range_pattern(old, new, sizeof((const char*[]){ __VA_ARGS__ }) / sizeof(const char*), (const char*[]){ __VA_ARGS__ })
 static uint64_t range_pattern(uint64_t old, uint64_t new, int c, const char* ranges[]) {
-    if (new < old) { fprintf(stderr, "error: our DFA can't travel backwards %llu -> %llu\n", old, new); abort(); }
+    if (new < old) {
+      fprintf(stderr, "error: our DFA can't travel backwards %llu -> %llu\n",
+          (long long unsigned) old,
+          (long long unsigned) new);
+      abort();
+    }
     assert(new - old < 15);
 
     for (int i = 0; i < c; i++) {
@@ -202,7 +207,12 @@ static uint64_t range_pattern(uint64_t old, uint64_t new, int c, const char* ran
 
 #define CHARS(old, new, ...) chars_pattern(old, new, __VA_ARGS__)
 static uint64_t chars_pattern(uint64_t old, uint64_t new, const char* str) {
-    if (new < old) { fprintf(stderr, "error: our DFA can't travel backwards %llu %llu\n", old, new); abort(); }
+    if (new < old) {
+      fprintf(stderr, "error: our DFA can't travel backwards %llu %llu\n",
+          (long long unsigned) old,
+          (long long unsigned) new);
+      abort();
+    }
     assert(new - old < 15);
 
     for (; *str; str++) {
@@ -389,7 +399,7 @@ int main(int argc, char** argv) {
 
         v[1] <<= 1;
 
-        fprintf(file, "    { 0x%016llx, 0x%016llx },\n", v[0], v[1]);
+        fprintf(file, "    { 0x%016llx, 0x%016llx },\n", (long long unsigned) v[0], (long long unsigned) v[1]);
     }
     fprintf(file, "};\n\n");
     /*fprintf(file, "static const uint8_t eq_classes[128] = {");
